requirements.txt
gradio>=3.40.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.23.0
pillow>=9.0.0
setuptools
